Move the actual GUI code to it's own subpackage, just invoke it in script
